modified

+++ app/src/main/java/hexlet/code/model/UrlCheck.java
@@ -16,10 +16,10 @@ public class UrlCheck {
     private String title;
     private String h1;
     private String description;
-    private int urlId;
+    private Long urlId;
     private LocalDateTime createdAt;
 
-    public UrlCheck(int statusCode, String title, String h1, String description, int urlId) {
+    public UrlCheck(int statusCode, String title, String h1, String description, Long urlId) {
         this.statusCode = statusCode;
         this.title = title;
         this.h1 = h1;
